Add TypeKey struct and fix IsArray type FullName
Introduced a new struct, TypeKey, to uniquely identify a Type and facilitate its usage in a dictionary as a key. Updated TypeExtensionMethods to make use of this new struct for its various caches, improving organization and access to these cached values. Added a new test method in FullNameTest.cs to verify cache functionality.